Share multithread utils between tests and benchmarks
