+++ benchmark/multithread.hpp
@@ -17,75 +17,9 @@
 #include <thread>
 #include <vector>
 
+#include "multithread_helpers.hpp"
+
 namespace umf_bench {
-
-template <typename Function>
-void parallel_exec(size_t threads_number, Function &&f) {
-    std::vector<std::thread> threads;
-    threads.reserve(threads_number);
-
-    for (size_t i = 0; i < threads_number; ++i) {
-        threads.emplace_back([&](size_t id) { f(id); }, i);
-    }
-
-    for (auto &t : threads) {
-        t.join();
-    }
-}
-
-class latch {
-  public:
-    latch(size_t desired) : counter(desired) {}
-
-    /* Returns true for the last thread arriving at the latch, false for all
-     * other threads. */
-    bool wait(std::unique_lock<std::mutex> &lock) {
-        counter--;
-        if (counter > 0) {
-            cv.wait(lock, [&] { return counter == 0; });
-            return false;
-        } else {
-            /*
-             * notify_call could be called outside of a lock
-             * (it would perform better) but drd complains
-             * in that case
-             */
-            cv.notify_all();
-            return true;
-        }
-    }
-
-  private:
-    std::condition_variable cv;
-    size_t counter = 0;
-};
-
-/* Implements multi-use barrier (latch). Once all threads arrive at the
- * latch, a new latch is allocated and used by all subsequent calls to
- * syncthreads. */
-struct syncthreads_barrier {
-    syncthreads_barrier(size_t num_threads) : num_threads(num_threads) {
-        mutex = std::shared_ptr<std::mutex>(new std::mutex);
-        current_latch = std::shared_ptr<latch>(new latch(num_threads));
-    }
-
-    syncthreads_barrier(const syncthreads_barrier &) = delete;
-    syncthreads_barrier &operator=(const syncthreads_barrier &) = delete;
-    syncthreads_barrier(syncthreads_barrier &&) = default;
-
-    void operator()() {
-        std::unique_lock<std::mutex> lock(*mutex);
-        auto l = current_latch;
-        if (l->wait(lock)) {
-            current_latch = std::shared_ptr<latch>(new latch(num_threads));
-        }
-    }
-
-  private:
-    size_t num_threads;
-    std::shared_ptr<std::mutex> mutex;
-    std::shared_ptr<latch> current_latch;
-};
 
 template <typename TimeUnit, typename F>
 typename TimeUnit::rep measure(F &&func) {
@@ -110,14 +44,12 @@ auto measure(size_t iterations, size_t concurrency, F &&run_workload) {
 
     for (size_t i = 0; i < iterations; i++) {
         std::vector<ResultsType> iteration_results(concurrency);
-        syncthreads_barrier syncthreads(concurrency);
-        parallel_exec(concurrency, [&](size_t id) {
+        umf_test::syncthreads_barrier syncthreads(concurrency);
+        umf_test::parallel_exec(concurrency, [&](size_t id) {
             syncthreads();
 
             iteration_results[id] =
                 measure<TimeUnit>([&]() { run_workload(id); });
-
-            syncthreads();
         });
 
         // skip the first 'warmup' iteration
