+++ test/supp/drd-umf_test-ipc.supp
@@ -1,0 +1,7 @@
+{
+   Conditional variable destruction false-positive
+   drd:CondErr
+   ...
+   fun:pthread_cond_destroy@*
+   ...
+}
